address review nits
